V 1.4

+++ application/modules/admin/views/pages/pages.php
@@ -16,17 +16,25 @@
 					<th>Delete</th>
 				</tr>
 			</thead>
-			<?php foreach($pages as $cont): ?>
-			<tbody>
-				<tr>
-					<td><?php echo $cont->page_id; ?></td>
-					<td><?php echo $cont->page_name; ?></td>
-					<td><?php echo $cont->display_name; ?></td>
-					<td><a href="<?php echo base_url(); ?>admin/pages/edit_page/<?php echo $cont->page_name; ?>"><span class="glyphicon glyphicon-pencil"></span></a></td>
-					<td><a href="<?php echo base_url(); ?>admin/deletePage/<?php echo $cont->page_name; ?>"><span class="glyphicon glyphicon-trash"></span></a></td>
-				</tr>
-			</tbody>			
-			<?php endforeach; ?>
+			<?php if(!empty($pages)): ?>
+				<?php foreach($pages as $cont): ?>
+				<tbody>
+					<tr>
+						<td><?php echo $cont->page_id; ?></td>
+						<td><?php echo $cont->page_name; ?></td>
+						<td><?php echo $cont->display_name; ?></td>
+						<td><a href="<?php echo base_url(); ?>admin/pages/edit_page/<?php echo $cont->page_name; ?>"><span class="glyphicon glyphicon-pencil"></span></a></td>
+						<td><a href="<?php echo base_url(); ?>admin/deletePage/<?php echo $cont->page_name; ?>"><span class="glyphicon glyphicon-trash"></span></a></td>
+					</tr>
+				</tbody>			
+				<?php endforeach; ?>
+			<?php else: ?>
+				<tbody>
+					<tr>
+						<td colspan="5" class="text-center">No Data Available</td>
+					</tr>
+				</tbody>
+			<?php endif; ?>
 		</table>
 	</div>
 </div>
